test(terraform#project): extract shared invoke helpers and restore full deploy coverage
