+++ e2e/src/smoke-tests/cdk-deploy.spec.ts
@@ -6,294 +6,18 @@ import { existsSync, readFileSync, rmSync, writeFileSync } from 'fs';
 import { ensureDirSync } from 'fs-extra';
 import { runCLI, tmpProjPath } from '../utils';
 import { join } from 'path';
-import { AwsClient } from 'aws4fetch';
 import { CloudFormation, StackStatus } from '@aws-sdk/client-cloudformation';
-import { Lambda } from '@aws-sdk/client-lambda';
-import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
 import { runSmokeTest } from './smoke-test';
-import { Client } from '@modelcontextprotocol/sdk/client/index.js';
-import { StreamableHTTPClientTransport } from '@modelcontextprotocol/sdk/client/streamableHttp.js';
-
-// @ts-expect-error no types for virtual module
-import { AgentCoreTrpcClient } from 'virtual:ts-template/ts/strands-agent/files/http/agent-core-trpc-client';
-
-/**
- * Helper function to create an AWS client with SigV4 signing
- */
-async function createAwsClient(service = 'execute-api'): Promise<AwsClient> {
-  const credentials = await fromNodeProviderChain()();
-  return new AwsClient({
-    accessKeyId: credentials.accessKeyId,
-    secretAccessKey: credentials.secretAccessKey,
-    sessionToken: credentials.sessionToken,
-    service,
-    region: process.env.AWS_REGION || 'us-west-2',
-  });
-}
-
-/**
- * Helper function to build Bedrock Agent Core Runtime URL
- */
-function buildAgentCoreUrl(arn: string): string {
-  const encodedArn = arn.replace(/:/g, '%3A').replace(/\//g, '%2F');
-  const region = process.env.AWS_REGION || 'us-west-2';
-  return `https://bedrock-agentcore.${region}.amazonaws.com/runtimes/${encodedArn}/invocations?qualifier=DEFAULT`;
-}
-
-/**
- * Helper function to invoke a tRPC API endpoint
- */
-async function invokeTrpcApi(apiUrl: string, apiName: string): Promise<void> {
-  const aws = await createAwsClient();
-  console.log(`Testing ${apiName} at ${apiUrl}`);
-  const input = encodeURIComponent(JSON.stringify({ message: 'test' }));
-  const response = await aws.fetch(`${apiUrl}echo?input=${input}`);
-  const data = await response.json();
-  console.log(`${apiName} response:`, data);
-  expect(data.result.data.result).toBe('test');
-}
-
-/**
- * Helper function to invoke a REST API endpoint (eg FastAPI or Smithy)
- */
-async function invokeRestApi(apiUrl: string, apiName: string): Promise<void> {
-  const aws = await createAwsClient();
-  console.log(`Testing ${apiName} at ${apiUrl}`);
-  const response = await aws.fetch(`${apiUrl}echo?message=test`);
-  const data = await response.json();
-  console.log(`${apiName} response:`, data);
-  expect(data.message).toBe('test');
-}
-
-/**
- * Helper function to invoke an MCP server via Bedrock Agent Core Runtime
- */
-async function invokeAgentCoreMcp(arn: string, mcpName: string): Promise<void> {
-  const aws = await createAwsClient('bedrock-agentcore');
-  console.log(`Testing ${mcpName} with ARN ${arn}`);
-
-  const mcpUrl = buildAgentCoreUrl(arn);
-
-  // Create MCP client with custom fetch
-  const transport = new StreamableHTTPClientTransport(new URL(mcpUrl), {
-    fetch: aws.fetch.bind(aws),
-  });
-
-  const client = new Client({
-    name: 'test-client',
-    version: '1.0.0',
-  });
-
-  await client.connect(transport);
-
-  try {
-    await client.ping();
-    console.log(`Pinged MCP server ${mcpName}`);
-  } finally {
-    await client.close();
-  }
-}
-
-/**
- * Helper function to invoke a Strands Agent via Bedrock Agent Core Runtime
- */
-async function invokeAgentCoreAgent(
-  arn: string,
-  agentName: string,
-  message = 'what is 3 + 5 - 2?',
-): Promise<string> {
-  const aws = await createAwsClient('bedrock-agentcore');
-  console.log(`Testing ${agentName} with ARN ${arn}`);
-
-  const agentUrl = buildAgentCoreUrl(arn);
-
-  const response = await aws.fetch(agentUrl, {
-    method: 'POST',
-    body: JSON.stringify({ message }),
-    headers: {
-      'Content-Type': 'application/json',
-      'X-Amzn-Bedrock-AgentCore-Runtime-Session-Id':
-        'abcdefghijklmnopqrstuvwxyz0123456789',
-    },
-  });
-
-  const chunks: { content: string }[] = [];
-  let buffer = '';
-  const reader = response.body.pipeThrough(new TextDecoderStream()).getReader();
-  while (reader) {
-    const { value, done } = await reader.read();
-    if (done) {
-      if (buffer.trim()) {
-        chunks.push(JSON.parse(buffer));
-      }
-      break;
-    }
-    buffer += value;
-    const lines = buffer.split('\n');
-    buffer = lines.pop() ?? '';
-    for (const line of lines) {
-      if (line.trim()) {
-        const chunk = JSON.parse(line);
-        chunks.push(chunk);
-        console.log(chunk);
-      }
-    }
-  }
-
-  console.log('Agent Response chunks:', chunks);
-
-  console.log(`${agentName} response status:`, response.status);
-  expect(response.status).toBe(200);
-  expect(chunks.length).toBeGreaterThan(0);
-  expect(chunks.every((c) => typeof c.content === 'string')).toBe(true);
-  console.log(`Successfully invoked ${agentName}`);
-  return chunks.map((c) => c.content).join('');
-}
-
-async function invokeTrpcAgentCoreAgent(
-  arn: string,
-  agentName: string,
-  message = 'what is 3 * 5 / 4?',
-): Promise<string> {
-  console.log(`Testing ${agentName} with ARN ${arn}`);
-
-  // Dogfood our vended client
-  const client = AgentCoreTrpcClient.withIamAuth({
-    agentRuntimeArn: arn,
-  });
-
-  // Wait for any data, or an error
-  const response = await new Promise<string>((resolve, reject) => {
-    let responseMessage = '';
-
-    // NB the trpc api is generated code so we don't have a type-safe trpc client here
-    (client.invoke as any).subscribe(
-      { message },
-      {
-        onData: (chunk: string) => {
-          console.log(chunk);
-          responseMessage += chunk;
-        },
-        onComplete: () => resolve(responseMessage),
-        onError: (error: any) => reject(error),
-      },
-    );
-  });
-  console.log('Agent Response', response);
-  expect(response).not.toHaveLength(0);
-  return response;
-}
-
-/**
- * Helper function to invoke an A2A server via Bedrock AgentCore Runtime.
- *
- * Uses the A2A SDK's `ClientFactory` with a SigV4-signing fetch for both the
- * agent card resolver and the JSON-RPC transport — this exercises the same
- * code path that `A2AAgent.clientFactory` uses inside generated agents.
- */
-async function invokeAgentCoreA2a(
-  arn: string,
-  agentName: string,
-): Promise<void> {
-  // Lazy imports so these SDKs are only required in tests that need them
-  const {
-    ClientFactory,
-    ClientFactoryOptions,
-    DefaultAgentCardResolver,
-    JsonRpcTransportFactory,
-  } = await import('@a2a-js/sdk/client');
-
-  console.log(`Testing ${agentName} with ARN ${arn}`);
-
-  const region = process.env.AWS_REGION || 'us-west-2';
-  const encodedArn = encodeURIComponent(arn);
-  // AgentCore mounts A2A at the /invocations/ root (trailing slash matters).
-  const baseUrl = `https://bedrock-agentcore.${region}.amazonaws.com/runtimes/${encodedArn}/invocations/`;
-  const sessionId = 'abcdefghijklmnopqrstuvwxyz0123456789';
-
-  const credentialsProvider = fromNodeProviderChain();
-  const sigv4Fetch: typeof fetch = async (input, init) => {
-    const credentials = await credentialsProvider();
-    const awsClient = new AwsClient({
-      accessKeyId: credentials.accessKeyId,
-      secretAccessKey: credentials.secretAccessKey,
-      sessionToken: credentials.sessionToken,
-      service: 'bedrock-agentcore',
-      region,
-    });
-    const headers = new Headers(init?.headers);
-    if (!headers.has('X-Amzn-Bedrock-AgentCore-Runtime-Session-Id')) {
-      headers.set('X-Amzn-Bedrock-AgentCore-Runtime-Session-Id', sessionId);
-    }
-    return awsClient.fetch(input, { ...init, headers });
-  };
-
-  // NB: `cardResolver` must be passed to the ClientFactory directly —
-  // `ClientFactoryOptions.createFrom` does not thread it through.
-  const factory = new ClientFactory({
-    ...ClientFactoryOptions.default,
-    transports: [new JsonRpcTransportFactory({ fetchImpl: sigv4Fetch })],
-    cardResolver: new DefaultAgentCardResolver({ fetchImpl: sigv4Fetch }),
-  });
-
-  const client = await factory.createFromUrl(baseUrl);
-  const card = await client.getAgentCard();
-  expect(typeof card.name).toBe('string');
-  console.log(`${agentName} agent card:`, card.name);
-
-  const stream = client.sendMessageStream({
-    message: {
-      kind: 'message',
-      role: 'user',
-      parts: [{ kind: 'text', text: 'hello' }],
-      messageId: crypto.randomUUID(),
-    },
-  });
-
-  let events = 0;
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  for await (const _event of stream) {
-    events++;
-    break;
-  }
-  expect(events).toBeGreaterThan(0);
-  console.log(`Successfully invoked ${agentName} (${events} events)`);
-}
-
-async function invokeLambda(arn: string, lambdaName: string) {
-  const lambda = new Lambda();
-  console.log('Invoking lambda', lambdaName, 'with arn', arn);
-  const res = await lambda.invoke({
-    FunctionName: arn,
-    Payload: JSON.stringify({}),
-  });
-  const payload = res.Payload
-    ? new TextDecoder().decode(res.Payload)
-    : undefined;
-  console.log('Status code', res.StatusCode);
-  console.log('Function error', res.FunctionError);
-  console.log('Payload', payload);
-  // Lambda returns StatusCode 200 even when the handler throws — the handler
-  // failure surfaces via FunctionError. Assert both to catch runtime errors
-  // like missing-module ImportErrors that only manifest in a deployed Lambda.
-  expect(res.StatusCode).toBe(200);
-  expect(res.FunctionError).toBeUndefined();
-  console.log(`Successfully invoked ${lambdaName}`);
-}
-
-/**
- * Helper function to check that the website has been deployed
- */
-async function pingWebsite(domain: string) {
-  console.log('Testing website with domain', domain);
-  const status = (await fetch(`https://${domain}/`)).status;
-  console.log('Website response status:', status);
-  expect(status).toBe(200);
-  expect(
-    await (await fetch(`https://${domain}/runtime-config.json`)).json(),
-  ).toHaveProperty('cognitoProps');
-  console.log('Successfully pinged website');
-}
+import {
+  invokeAgentCoreA2a,
+  invokeAgentCoreAgent,
+  invokeAgentCoreMcp,
+  invokeLambda,
+  invokeRestApi,
+  invokeTrpcAgentCoreAgent,
+  invokeTrpcApi,
+  pingWebsite,
+} from './deploy-invocations';
 
 /**
  * Delete any CloudFormation stacks matching the test run prefix that were
