+++ e2e/src/smoke-tests/terraform-deploy.spec.ts
@@ -7,234 +7,18 @@ import { existsSync, readFileSync, rmSync, writeFileSync } from 'fs';
 import { ensureDirSync } from 'fs-extra';
 import fastGlob from 'fast-glob';
 import { join } from 'path';
-import { AwsClient } from 'aws4fetch';
-import { Lambda } from '@aws-sdk/client-lambda';
-import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
-import { Client } from '@modelcontextprotocol/sdk/client/index.js';
-import { StreamableHTTPClientTransport } from '@modelcontextprotocol/sdk/client/streamableHttp.js';
 
 import { runCLI, tmpProjPath } from '../utils';
 import { runTerraformSmokeTest } from './terraform-smoke-test';
-
-/**
- * Helper function to create an AWS client with SigV4 signing
- */
-async function createAwsClient(service = 'execute-api'): Promise<AwsClient> {
-  const credentials = await fromNodeProviderChain()();
-  return new AwsClient({
-    accessKeyId: credentials.accessKeyId,
-    secretAccessKey: credentials.secretAccessKey,
-    sessionToken: credentials.sessionToken,
-    service,
-    region: process.env.AWS_REGION || 'us-west-2',
-  });
-}
-
-/**
- * Helper function to build Bedrock Agent Core Runtime URL
- */
-function buildAgentCoreUrl(arn: string): string {
-  const encodedArn = arn.replace(/:/g, '%3A').replace(/\//g, '%2F');
-  const region = process.env.AWS_REGION || 'us-west-2';
-  return `https://bedrock-agentcore.${region}.amazonaws.com/runtimes/${encodedArn}/invocations?qualifier=DEFAULT`;
-}
-
-/**
- * Helper function to invoke a tRPC API endpoint (reached via the terraform
- * `stage_invoke_url` output, which does NOT include a trailing slash).
- */
-// Normalise API URLs — REST stages come back as `...amazonaws.com/prod`
-// while HTTP `$default` stages come back as `...amazonaws.com/`. Appending
-// `/echo` naively would produce `...amazonaws.com//echo` for the HTTP case
-// and AWS SigV4 rejects that as an auth mismatch ("Forbidden").
-const normalizeApiUrl = (url: string): string => url.replace(/\/$/, '');
-
-async function invokeTrpcApi(apiUrl: string, apiName: string): Promise<void> {
-  const aws = await createAwsClient();
-  console.log(`Testing ${apiName} at ${apiUrl}`);
-  const input = encodeURIComponent(JSON.stringify({ message: 'test' }));
-  const response = await aws.fetch(
-    `${normalizeApiUrl(apiUrl)}/echo?input=${input}`,
-  );
-  const data = await response.json();
-  console.log(`${apiName} response:`, data);
-  expect(data.result.data.result).toBe('test');
-}
-
-async function invokeRestApi(apiUrl: string, apiName: string): Promise<void> {
-  const aws = await createAwsClient();
-  console.log(`Testing ${apiName} at ${apiUrl}`);
-  const response = await aws.fetch(
-    `${normalizeApiUrl(apiUrl)}/echo?message=test`,
-  );
-  const data = await response.json();
-  console.log(`${apiName} response:`, data);
-  expect(data.message).toBe('test');
-}
-
-async function invokeAgentCoreMcp(arn: string, mcpName: string): Promise<void> {
-  const aws = await createAwsClient('bedrock-agentcore');
-  console.log(`Testing ${mcpName} with ARN ${arn}`);
-
-  const transport = new StreamableHTTPClientTransport(
-    new URL(buildAgentCoreUrl(arn)),
-    {
-      fetch: aws.fetch.bind(aws),
-    },
-  );
-
-  const client = new Client({ name: 'test-client', version: '1.0.0' });
-  await client.connect(transport);
-  try {
-    await client.ping();
-    console.log(`Pinged MCP server ${mcpName}`);
-  } finally {
-    await client.close();
-  }
-}
-
-async function invokeAgentCoreAgent(
-  arn: string,
-  agentName: string,
-  message = 'what is 3 + 5 - 2?',
-): Promise<string> {
-  const aws = await createAwsClient('bedrock-agentcore');
-  console.log(`Testing ${agentName} with ARN ${arn}`);
-
-  const response = await aws.fetch(buildAgentCoreUrl(arn), {
-    method: 'POST',
-    body: JSON.stringify({ message }),
-    headers: {
-      'Content-Type': 'application/json',
-      'X-Amzn-Bedrock-AgentCore-Runtime-Session-Id':
-        'abcdefghijklmnopqrstuvwxyz0123456789',
-    },
-  });
-
-  const chunks: { content: string }[] = [];
-  let buffer = '';
-  const reader = response.body.pipeThrough(new TextDecoderStream()).getReader();
-  while (reader) {
-    const { value, done } = await reader.read();
-    if (done) {
-      if (buffer.trim()) {
-        chunks.push(JSON.parse(buffer));
-      }
-      break;
-    }
-    buffer += value;
-    const lines = buffer.split('\n');
-    buffer = lines.pop() ?? '';
-    for (const line of lines) {
-      if (line.trim()) {
-        const chunk = JSON.parse(line);
-        chunks.push(chunk);
-        console.log(chunk);
-      }
-    }
-  }
-
-  console.log(`${agentName} response status:`, response.status);
-  expect(response.status).toBe(200);
-  expect(chunks.length).toBeGreaterThan(0);
-  expect(chunks.every((c) => typeof c.content === 'string')).toBe(true);
-  console.log(`Successfully invoked ${agentName}`);
-  return chunks.map((c) => c.content).join('');
-}
-
-async function invokeAgentCoreA2a(
-  arn: string,
-  agentName: string,
-): Promise<void> {
-  const {
-    ClientFactory,
-    ClientFactoryOptions,
-    DefaultAgentCardResolver,
-    JsonRpcTransportFactory,
-  } = await import('@a2a-js/sdk/client');
-
-  console.log(`Testing ${agentName} with ARN ${arn}`);
-
-  const region = process.env.AWS_REGION || 'us-west-2';
-  const encodedArn = encodeURIComponent(arn);
-  const baseUrl = `https://bedrock-agentcore.${region}.amazonaws.com/runtimes/${encodedArn}/invocations/`;
-  const sessionId = 'abcdefghijklmnopqrstuvwxyz0123456789';
-
-  const credentialsProvider = fromNodeProviderChain();
-  const sigv4Fetch: typeof fetch = async (input, init) => {
-    const credentials = await credentialsProvider();
-    const awsClient = new AwsClient({
-      accessKeyId: credentials.accessKeyId,
-      secretAccessKey: credentials.secretAccessKey,
-      sessionToken: credentials.sessionToken,
-      service: 'bedrock-agentcore',
-      region,
-    });
-    const headers = new Headers(init?.headers);
-    if (!headers.has('X-Amzn-Bedrock-AgentCore-Runtime-Session-Id')) {
-      headers.set('X-Amzn-Bedrock-AgentCore-Runtime-Session-Id', sessionId);
-    }
-    return awsClient.fetch(input, { ...init, headers });
-  };
-
-  const factory = new ClientFactory({
-    ...ClientFactoryOptions.default,
-    transports: [new JsonRpcTransportFactory({ fetchImpl: sigv4Fetch })],
-    cardResolver: new DefaultAgentCardResolver({ fetchImpl: sigv4Fetch }),
-  });
-
-  const client = await factory.createFromUrl(baseUrl);
-  const card = await client.getAgentCard();
-  expect(typeof card.name).toBe('string');
-  console.log(`${agentName} agent card:`, card.name);
-
-  const stream = client.sendMessageStream({
-    message: {
-      kind: 'message',
-      role: 'user',
-      parts: [{ kind: 'text', text: 'hello' }],
-      messageId: crypto.randomUUID(),
-    },
-  });
-
-  let events = 0;
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  for await (const _event of stream) {
-    events++;
-    break;
-  }
-  expect(events).toBeGreaterThan(0);
-  console.log(`Successfully invoked ${agentName} (${events} events)`);
-}
-
-async function invokeLambda(arn: string, lambdaName: string) {
-  const lambda = new Lambda();
-  console.log('Invoking lambda', lambdaName, 'with arn', arn);
-  const res = await lambda.invoke({
-    FunctionName: arn,
-    Payload: JSON.stringify({}),
-  });
-  const payload = res.Payload
-    ? new TextDecoder().decode(res.Payload)
-    : undefined;
-  console.log('Status code', res.StatusCode);
-  console.log('Function error', res.FunctionError);
-  console.log('Payload', payload);
-  expect(res.StatusCode).toBe(200);
-  expect(res.FunctionError).toBeUndefined();
-  console.log(`Successfully invoked ${lambdaName}`);
-}
-
-async function pingWebsite(domain: string) {
-  console.log('Testing website with domain', domain);
-  const status = (await fetch(`https://${domain}/`)).status;
-  console.log('Website response status:', status);
-  expect(status).toBe(200);
-  expect(
-    await (await fetch(`https://${domain}/runtime-config.json`)).json(),
-  ).toHaveProperty('cognitoProps');
-  console.log('Successfully pinged website');
-}
+import {
+  invokeAgentCoreA2a,
+  invokeAgentCoreAgent,
+  invokeAgentCoreMcp,
+  invokeLambda,
+  invokeRestApi,
+  invokeTrpcApi,
+  pingWebsite,
+} from './deploy-invocations';
 
 /**
  * Read Terraform outputs as JSON from the infra project.
@@ -272,72 +56,43 @@ describe('smoke test - terraform-deploy', () => {
   it('should generate, deploy, exercise and destroy', async () => {
     const { opts } = await runTerraformSmokeTest(targetDir, pkgMgr);
 
-    // 8-digit alphanumeric test run id, used to make per-test state keys and
-    // per-test IAM policy names unique across concurrent PRs.
+    // 8-digit alphanumeric test run id — used to make per-test state keys
+    // and every hardcoded resource name unique so concurrent PR runs and
+    // retries after a failed destroy don't collide on `EntityAlreadyExists`.
     const testRunId = Math.random().toString(36).substring(2, 10);
 
-    // Overwrite the scaffolded main.tf with our wiring template. This is the
-    // terraform analogue of the cdk-deploy `application-stack.ts.template`.
-    const templateContent = readFileSync(
+    // Overwrite the scaffolded main.tf with our wiring template (the
+    // Terraform analogue of the `application-stack.ts.template` the
+    // cdk-deploy test uses).
+    const mainTfTemplate = readFileSync(
       join(__dirname, '../files/terraform-deploy/main.tf.template'),
       'utf-8',
     );
-    const mainContent = templateContent.replace(
-      /<% TEST_RUN_ID %>/g,
-      testRunId,
+    writeFileSync(
+      `${opts.cwd}/packages/infra/src/main.tf`,
+      mainTfTemplate.replace(/<% TEST_RUN_ID %>/g, testRunId),
     );
-    writeFileSync(`${opts.cwd}/packages/infra/src/main.tf`, mainContent);
 
-    // Patch the infra project's init target to use a per-run state key.
-    // `plan` depends on `init`, so every `nx apply` / `nx destroy` invocation
-    // re-runs init; hard-coding the override in project.json is the least
-    // fragile way to ensure concurrent test runs don't collide on state.
+    // Patch the infra project targets:
+    // - `init`: use a per-run state key so concurrent PR runs don't share state
+    // - `destroy`: auto-approve for unattended teardown
     const infraProjectJsonPath = `${opts.cwd}/packages/infra/project.json`;
     const infraProjectJson = JSON.parse(
       readFileSync(infraProjectJsonPath, 'utf-8'),
     );
-    const initDevCmd: string =
-      infraProjectJson.targets.init.configurations.dev.command;
     infraProjectJson.targets.init.configurations.dev.command =
-      initDevCmd.replace(
+      infraProjectJson.targets.init.configurations.dev.command.replace(
         /e2e-test-infra\/dev\/terraform\.tfstate/,
         `e2e-test-infra-${testRunId}/dev/terraform.tfstate`,
       );
-
-    // Force -parallelism=1 on plan, apply and destroy. The generated
-    // appconfig module reads the runtime-config directory via
-    // `fileset(config_dir, "*.json")` which races against parallel
-    // `add_*_runtime_config` external data sources that write namespace
-    // JSON files during the same walk — surfacing as
-    // "function returned an inconsistent result".
-    const planCommands: string[] =
-      infraProjectJson.targets.plan.configurations.dev.commands;
-    infraProjectJson.targets.plan.configurations.dev.commands =
-      planCommands.map((cmd) =>
-        cmd.startsWith('terraform plan') ? `${cmd} -parallelism=1` : cmd,
-      );
-    // `terraform apply` takes `-parallelism=N` before the plan file
-    // positional argument; splice the flag in right after `terraform apply`.
-    // `-refresh=false` stops terraform from re-running `data.external`
-    // sources during apply — those are the ones that re-mutate the
-    // runtime-config JSON files and trigger the appconfig `file()` race.
-    infraProjectJson.targets.apply.configurations.dev.command =
-      infraProjectJson.targets.apply.configurations.dev.command.replace(
-        /^terraform apply\b/,
-        'terraform apply -parallelism=1 -refresh=false',
-      );
-    // Destroy doesn't hit the same fileset race (no new files get written
-    // during teardown), so leave parallelism at the default to keep the
-    // ~20min destroy phase from ballooning.
-
     writeFileSync(
       infraProjectJsonPath,
       JSON.stringify(infraProjectJson, null, 2),
     );
 
-    // Enable auto-sync so generated files (license headers on the new main.tf,
-    // ts path aliases etc.) don't block the `apply` / `destroy` pipeline with
-    // a fatal "workspace is out of sync" error.
+    // Enable auto-sync so generated files (license headers on the new
+    // main.tf, ts path aliases etc.) don't block the `apply` / `destroy`
+    // pipeline with a fatal "workspace is out of sync" error.
     const nxJsonPath = `${opts.cwd}/nx.json`;
     const nxJson = JSON.parse(readFileSync(nxJsonPath, 'utf-8'));
     nxJson.sync = { ...(nxJson.sync ?? {}), applyChanges: true };
@@ -355,39 +110,13 @@ describe('smoke test - terraform-deploy', () => {
       ),
     );
 
-    // Replace the generated appconfig module's racy `fileset`/`file`
-    // discovery with a static namespace list. The shipped implementation
-    // reads the `runtime-config` JSON files lazily at evaluation time;
-    // sibling `data.external.updated_config` modules keep mutating those
-    // files mid-walk, so terraform's internal consistency check re-evaluates
-    // the local and sees different content on the second pass — surfacing as
-    // "Call to function 'file' failed: function returned an inconsistent
-    // result". Using a hardcoded map of namespaces with empty objects keeps
-    // the ConfigurationProfile / HostedConfigurationVersion / Deployment
-    // resources that downstream agents/MCP servers need (otherwise the agent
-    // runtime crashes at startup with "ConfigurationProfile not found")
-    // without reading anything off disk.
-    const appconfigTfPath = `${opts.cwd}/packages/common/terraform/src/core/runtime-config/appconfig/appconfig.tf`;
-    writeFileSync(
-      appconfigTfPath,
-      readFileSync(appconfigTfPath, 'utf-8').replace(
-        /# Read all namespace JSON files[\s\S]*?\n\}/m,
-        `locals {
-  namespaces = {
-    connection = {}
-    agentcore  = {}
-  }
-}`,
-      ),
-    );
-
     // Inject `testRunId` into every hardcoded resource name in the generated
     // app modules. The generators produce module-internal names like
     // `api_name = "MyApi"`, `function_name = "MyApiHandler"`,
-    // `application_name = "MyAgent-runtime-config"`, etc. Without a per-run
-    // suffix, concurrent test runs (or a failed run that left leftovers)
-    // collide on "EntityAlreadyExists" errors the next time around. We mirror
-    // the CDK `testRunId` approach by suffixing every hardcoded name.
+    // `application_name = "MyAgent-runtime-config"` etc. — without a
+    // per-run suffix, concurrent test runs (or a retry after a failed
+    // destroy) collide on "EntityAlreadyExists". We mirror the CDK
+    // `testRunId` approach by suffixing every hardcoded name here.
     const rewriteNamesInFile = (
       path: string,
       rewrites: [RegExp, string][],
@@ -418,7 +147,7 @@ describe('smoke test - terraform-deploy', () => {
     }
 
     // API modules (REST + HTTP): suffix `api_name` and the hardcoded
-    // per-api Lambda/IAM/log-group names.
+    // per-api Lambda/IAM/log-group/KMS/S3-bucket names.
     for (const tfPath of fastGlob.sync(
       `${opts.cwd}/packages/common/terraform/src/app/apis/*/*.tf`,
     )) {
@@ -468,38 +197,15 @@ describe('smoke test - terraform-deploy', () => {
     }
 
     try {
-      // Apply sync one more time so the newly-written main.tf and patched
-      // project.json get their license headers. Nx in non-interactive mode
-      // won't auto-apply sync changes during `apply` — it aborts with a fatal
-      // "workspace is out of sync" error instead.
-      await runCLI(`sync`, opts);
-
-      // One-time remote-state bootstrap (idempotent — the target re-uses any
-      // existing tfstate already in the per-account/region bucket).
+      // One-time remote-state bootstrap (idempotent — the target re-uses
+      // any existing tfstate already in the per-account/region bucket).
       await runCLI(`bootstrap infra --output-style=stream`, opts);
 
-      // Pre-seed the runtime-config directory with all namespace JSON files
-      // the generated `add_*_runtime_config` modules will write during apply.
-      // Without this the appconfig module's `fileset(config_dir, "*.json")`
-      // sees the directory change between plan-time and apply-time evaluations
-      // and fails with "function returned an inconsistent result" — the
-      // external data sources that write these files run *during* the graph
-      // walk, after fileset has already been evaluated once.
-      const runtimeConfigDir = `${opts.cwd}/dist/packages/common/terraform/runtime-config`;
-      ensureDirSync(runtimeConfigDir);
-      for (const ns of ['connection', 'agentcore']) {
-        const nsPath = `${runtimeConfigDir}/${ns}.json`;
-        if (!existsSync(nsPath)) {
-          writeFileSync(nsPath, '{}');
-        }
-      }
-
-      // Plan + apply end to end. The plan target has been patched to pass
-      // `-parallelism=1` so the generated appconfig module's `fileset(...)`
-      // doesn't race against the parallel runtime-config writers. Combined
-      // with the cross-module `depends_on` chain we added in main.tf, every
-      // agent/MCP module's appconfig reads the runtime-config directory only
-      // after all writers have stabilised.
+      // Plan + apply end to end. The generated appconfig/entry modules
+      // now write per-entry `local_file` resources into
+      // `dist/…/runtime-config/entries/<namespace>/`, and the appconfig
+      // module aggregates those via a `null_resource` at apply time —
+      // so there's no longer a consistency race across the graph walk.
       await runCLI(`apply infra --output-style=stream`, opts);
 
       const outputs = readTerraformOutputs(opts.cwd);
@@ -523,9 +229,12 @@ describe('smoke test - terraform-deploy', () => {
       );
       await invokeAgentCoreMcp(outputs.py_mcp_server_arn, 'Python MCP Server');
 
-      // NOTE: The HTTP strands agents (`strands_agent` / `ts_strands_agent`)
-      // are scaffolded via the generator matrix but omitted from the deploy
-      // — see main.tf.template for why.
+      // Strands agents — direct invocation
+      await invokeAgentCoreAgent(outputs.strands_agent_arn, 'Strands Agent');
+      await invokeAgentCoreAgent(
+        outputs.ts_strands_agent_arn,
+        'TypeScript Strands Agent',
+      );
 
       // A2A (direct invocation via A2A SDK)
       await invokeAgentCoreA2a(
@@ -534,22 +243,34 @@ describe('smoke test - terraform-deploy', () => {
       );
       await invokeAgentCoreA2a(outputs.py_a2a_agent_arn, 'Python A2A Agent');
 
-      // NOTE: The A2A-via-delegate-tool tests from the CDK smoke test rely on
-      // AppConfig runtime-config namespaces to carry downstream agent ARNs
-      // from terraform into the host agents' environment. The generated
-      // appconfig module reads those namespace JSON files lazily via
-      // `fileset`/`file` at evaluation time, which races against the
-      // `data.external.updated_config` writers elsewhere in the graph —
-      // surfacing as "Call to function 'file' failed: function returned an
-      // inconsistent result". We strip the namespace bindings to unblock
-      // `terraform apply`; that drops A2A delegate-tool coverage but the
-      // direct A2A + MCP + agent invocations below still exercise the full
-      // AgentCore runtime + SigV4 + A2A SDK code path.
+      // A2A-via-delegate-tool — mirrors the CDK smoke test. The HTTP host
+      // agents resolve downstream agent ARNs from AppConfig; terraform's
+      // generated appconfig module now deep-merges leaf entry files into
+      // the two well-known namespaces at apply time, so the host agents
+      // can find their connected A2A targets end-to-end.
+      await invokeAgentCoreAgent(
+        outputs.ts_strands_agent_arn,
+        'TS Agent -> TS A2A (via askMyTsA2aAgent)',
+        'Use the askMyTsA2aAgent tool to ask the remote agent what 5 * 4 is. Return just the answer.',
+      );
+      await invokeAgentCoreAgent(
+        outputs.ts_strands_agent_arn,
+        'TS Agent -> PY A2A (via askMyPyA2aAgent)',
+        'Use the askMyPyA2aAgent tool to ask the remote agent what 11 + 2 is. Return just the answer.',
+      );
+      await invokeAgentCoreAgent(
+        outputs.strands_agent_arn,
+        'PY Agent -> TS A2A (via ask_my_ts_a2a_agent)',
+        'Use the ask_my_ts_a2a_agent tool to ask the remote agent what 9 - 3 is. Return just the answer.',
+      );
+      await invokeAgentCoreAgent(
+        outputs.strands_agent_arn,
+        'PY Agent -> PY A2A (via ask_my_py_a2a_agent)',
+        'Use the ask_my_py_a2a_agent tool to ask the remote agent what 7 + 8 is. Return just the answer.',
+      );
 
-      // Lambda functions. We skip the Python lambda — see the note in
-      // main.tf.template; the py-project aggregator bundles all agent deps
-      // and exceeds the Lambda direct-upload limit until the Terraform
-      // py#lambda-function module gains S3-backed upload support.
+      // Lambda functions
+      await invokeLambda(outputs.py_function_arn, 'Python Function');
       await invokeLambda(outputs.ts_function_arn, 'TypeScript Function');
 
       // Website
